Add a convenient Makefile.PL
This is going to provide a significant speedup
for the travis builds but could also help anyone
to come with a quick hack without the need of installing
the whole Dist::Zilla stack.

The Makefile.PL is ignored when producing a build and a fresh
one is always generated.

Makefile.PL auto-updated by dist.ini